Motion, custom entropy and reshaping
Implemented a checkbox to give users the option to not perform motion regression.

Implemented calculation of head motion parameters when entropy maps are calculated and motion regression box is ticked.

Implemented option to create your own custom entropy code.

Changed time scale input: now you can give the exact scale you would like, instead of number of scales (so 5 would originally calculate 1 to 5. Now it will just calculate 5. Can also do 1,5 to calculate 1 and 5).

The data was reshaped to reduce the number of for-loops from 4 to 2.
Vectorization and indexing was also used to reduce the number of if-statements, and reduce the amount of code within the for-loops.

For multiple time scales, the preprocessing is now no longer repeated for each scale.