Testing and fixing new code implementation
Tested code for fApEn without motion regression and frequency filtering for: m = 2, r = 0.25, tau = 1 and scale 1 to 5. Masking threshold: 0.8, and first 5 volumes were discarded.

Based on the above testing, mistakes in the new code were found and rectified.
1) The indexing attempted for the normalization did not work as thought. Now, no indexing is used and normalization is instead applied on all time-series.
2) Changed entropy calculation for-loop to include within the loop extracting the data out of the input image file. Before, the data extraction step was using 'i = length(indices)' from the previous for-loop.

For removal of regressors, the if-statement is put outside of the for-loop, as it only needs to be evaluated once.